added debug, made createGraph more generic, changed Int8 to Int16 for testing

+++ src/DCB.hs
@@ -32,12 +32,12 @@ type Matrix r e = Array r DIM2 e
 
 type Attr  = Matrix A.U Double
 -- | Adjacency-Matrix
-type Adj   = Matrix A.U Int8
+type Adj   = Matrix A.U Int16
 
 -- | Matrix of constraints
 
 --TODO: Haddoc!
-type Constraints = (Vector A.U Int8, Matrix A.U Double)
+type Constraints = (Vector A.U Int16, Matrix A.U Double)
 -- | A vector of weights indicating how much divergence is allowed in which dimension
 type MaxDivergence = Vector A.U Double
 -- | Make this special Scalar explicitly visible
@@ -135,7 +135,7 @@ constraint attr div req (_, (fulfill, constr), _) newNode =
                  0 -> min (f sh) (attr!sh)
                  1 -> max (f sh) (attr!sh)
         constrNew = A.computeUnboxedS $A.traverse constr id updateConstr
-        fulfillNew = A.zipWith (\i b -> if i == 1 && b then 1::Int8 else 0::Int8) fulfill
+        fulfillNew = A.zipWith (\i b -> if i == 1 && b then 1::Int16 else 0::Int16) fulfill
                 $A.zipWith (\thediv dist -> abs dist <= thediv) div $A.foldS (-) 0 constrNew
         nrHit = A.foldAllS (+) (0::Int) $A.map fromIntegral fulfillNew
     in if nrHit >= req then Just (A.computeS fulfillNew, constrNew) else Nothing
